solution to the data type not extending issue

+++ Cousera/AlgorithmicToolBox/01_maximumPairProduct.cpp
@@ -3,8 +3,8 @@
 
 using namespace std;
 
-int getMaxPairwiseProduct(vector<int> arr){
-    int maxValue = 0;
+int getMaxPairwiseProduct(vector<long long> arr){
+    long long maxValue = 0;
     for (int i = 0; i < arr.size(); i++)
         for (int j = i+1; j < arr.size(); j++)
             if(arr[i]*arr[j] > maxValue) maxValue = arr[i]*arr[j];
@@ -16,7 +16,7 @@ int main(){
     int n;
     cin >> n;
 
-    vector<int> arr(n);
+    vector<long long> arr(n);
     for (int i = 0; i < n; i++)
     {
         cin >> arr[i];
